+++ CHANGELOG.md
@@ -13,10 +13,11 @@ The format is based on **[Keep a Changelog](https://keepachangelog.com/en/1.1.0/
   - (placeholder)
 
 - **Changed**
-  - (placeholder)
+  - Improved batching of change notifications.
+  - Tests now (need to) include notification flushes <code>await act(async ()=>{});</code>
 
 - **Fixed**
-  - (placeholder)
+  - Singleton removal for the store in Provider.
 
 - **Security**
   - (placeholder)